- Removed all GEOPM_ERROR_RUNTIME uses in TreeCommunicator. - Removed two exception classes used only in the TreeCommunicator and Controller. - Replaced these exception classes with a special error number an the geopm::Exception class. - Don't depend on errno being zero. - Add an exception for not implemented errors. - Make sure that Exception class member m_err is never zero. - Fix for string format of off_t on 32 bit platform. - Fix for int to size_t comparison in test.

+++ test/PlatformImpTest.cpp
@@ -320,7 +320,7 @@ TEST_F(PlatformImpTest, write_msr_whitelist)
 
     size = strlen(key_buf);
     //Compare with size + 1 due to null character at end of file
-    EXPECT_TRUE((int)(val.seekg(0, std::ifstream::end).tellg()) == (size + 1));
+    EXPECT_TRUE((size_t)(val.seekg(0, std::ifstream::end).tellg()) == (size + 1));
     val.seekg(0, std::ifstream::beg);
 
     val_buf = (char*)malloc(size);
